requirements.txt
PyPDF2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
langchain-community
streamlit